Delegate Req.Steps.checksum,verify_checksum to Req.Checksum
